backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.24.1
aiofiles==23.2.1
supabase==2.0.2
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.0
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
lightgbm==4.1.0
prophet==1.1.5
openai==1.3.7
PyJWT==2.8.0
python-docx==1.1.0
reportlab==4.0.7
jinja2==3.1.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
aiohttp==3.9.1
beautifulsoup4==4.12.2